Added swinging action
Swinging action was added to the Swinging class. The player rotates
in a circle about the top of a rope. Players can also grab different
ropes of different sizes. The rope will rotate with the player as
expected. Upon ending the Swinging state, the player returns the
rope to its original size.